Policies (#22)
* Scaffold policies.

* Create user with necessary permissions.

* Test user crud permissions.

* Policies (#14)

* Sanity tests for client access

* Update ajax urls

* use template strings for string  concatenation

* Use re-introduce app root in ajax urls

* rearrange seeding order

* Refactor alter table add column with foreign key

* Prefer sqlite

* recitify travis set up

* fix user test method names

* Rename consumer.

* Add ide helper.

* Implement more policies.

* Minor code improvements.

* Implement, tests more policies.

* Numerous improvements.

* Add user auth related actions.

* Change function names to camel case.

* Change to no-content response.

* Add confirm email.

* Re-arrange in asc order.

* Minor improvements.

* Implement user auth related tests.

- Also test api client-credentials auth.

* Add actingAsClient to laravel/passport.

https://github.com/laravel/passport/releases/tag/v7.5.0

* Seed modules and permissions.

* Fix users route param pattern.

* Add relations to user response.

* Document user related actions.

* Policies (#18)

* Sanity tests for client access

* Update ajax urls

* use template strings for string  concatenation

* Use re-introduce app root in ajax urls

* rearrange seeding order

* Refactor alter table add column with foreign key

* Prefer sqlite

* recitify travis set up

* fix user test method names

* Seed created_at and updated_at columns for
clients.

* Associate seed password client with seed user

* Switch swagger documentation to YAML.

* Remove creator relation.

* Additional api usage documentation.

* Policies (#21)

* Sanity tests for client access

* Update ajax urls

* use template strings for string  concatenation

* Use re-introduce app root in ajax urls

* rearrange seeding order

* Refactor alter table add column with foreign key

* Prefer sqlite

* recitify travis set up

* fix user test method names

* Seed created_at and updated_at columns for
clients.

* Associate seed password client with seed user

* fixes #20

* Temporarily hide unsused routes

* documentation

* Revert "documentation"

This reverts commit d1f2012ec3f988b8d7c250561e45685b8e3c5af1.

* Updated local to latest doc file

* Improve api usage documentation.

+++ app/Http/Controllers/ModuleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Module;
+use App\Models\Permission;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -21,10 +22,14 @@ class ModuleController extends Controller
     /**
      * Get modules.
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
+        $this->authorize('viewAny', [Module::class]);
+
         $modules = Module::withTrashed()->get();
 
         return response(['modules' => $modules]);
@@ -35,10 +40,15 @@ class ModuleController extends Controller
      *
      * @param \Illuminate\Http\Request $request
      *
+     * @throws ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
+        $this->authorize('create', [Module::class]);
+
         $validator = Validator::make($request->all(), [
             'name' => 'required|max:20',
             'description' => 'sometimes|max:25',
@@ -73,10 +83,14 @@ class ModuleController extends Controller
      *
      * @param string $name
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function show($name)
     {
+        $this->authorize('view', [Module::class]);
+
         $module = Module::withTrashed()->findOrFail($name);
 
         return response($module);
@@ -88,10 +102,15 @@ class ModuleController extends Controller
      * @param \Illuminate\Http\Request $request
      * @param string                   $name
      *
+     * @throws ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $name)
     {
+        $this->authorize('update', [Module::class]);
+
         $module = Module::findOrFail($name);
 
         $this->validate($request, [
@@ -111,10 +130,15 @@ class ModuleController extends Controller
      *
      * @param string $name
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     * @throws \Exception
+     *
      * @return \Illuminate\Http\Response
      */
     public function revoke($name)
     {
+        $this->authorize('soft-delete', [Module::class]);
+
         $module = Module::findOrFail($name);
 
         $module->delete();
@@ -129,10 +153,14 @@ class ModuleController extends Controller
      *
      * @param string $name
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function restore($name)
     {
+        $this->authorize('restore', [Module::class]);
+
         $module = Module::onlyTrashed()->findOrFail($name);
 
         $module->restore();
@@ -147,11 +175,25 @@ class ModuleController extends Controller
      *
      * @param string $name
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function destroy($name)
     {
+        $this->authorize('force-delete', [Module::class]);
+
         $module = Module::onlyTrashed()->findOrFail($name);
+
+        // ...
+
+        $dependants = Permission::where('module_name', $name)->count();
+
+        if ($dependants) {
+            return response(['message' => "Can't delete non-orphaned module."], 400);
+        }
+
+        // ...
 
         $module->forceDelete();
