+++ app/Http/Controllers/RoleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Role;
+use App\Models\User;
 use App\Models\Permission;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -22,10 +23,14 @@ class RoleController extends Controller
     /**
      * Get roles.
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
+        $this->authorize('viewAny', [Role::class]);
+
         $user = Auth::guard('api')->user();
 
         $roles = Role::onlyRelated($user)->withTrashed()->get();
@@ -38,10 +43,15 @@ class RoleController extends Controller
      *
      * @param \Illuminate\Http\Request $request
      *
+     * @throws ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
+        $this->authorize('create', [Role::class]);
+
         $this->validate($request, [
             'name' => 'required|max:25',
             'description' => 'sometimes|max:50',
@@ -52,7 +62,6 @@ class RoleController extends Controller
         $role = new Role();
         $role->name = $request->name;
         $role->description = $request->description;
-        $role->creator()->associate($user);
         $role->facility()->associate($user->facility);
         $role->save();
 
@@ -66,10 +75,14 @@ class RoleController extends Controller
      *
      * @param string $id
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
+        $this->authorize('view', [Role::class, $id]);
+
         $user = Auth::guard('api')->user();
 
         $role = Role::onlyRelated($user)->withTrashed()->findOrFail($id);
@@ -83,10 +96,15 @@ class RoleController extends Controller
      * @param \Illuminate\Http\Request $request
      * @param string                   $id
      *
+     * @throws ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('update', [Role::class]);
+
         $user = Auth::guard('api')->user();
 
         $role = Role::onlyRelated($user)->findOrFail($id);
@@ -110,10 +128,14 @@ class RoleController extends Controller
      *
      * @param string $id
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function revoke($id)
     {
+        $this->authorize('softDelete', [Role::class]);
+
         $user = Auth::guard('api')->user();
 
         $role = Role::onlyRelated($user)->findOrFail($id);
@@ -130,10 +152,14 @@ class RoleController extends Controller
      *
      * @param string $id
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function restore($id)
     {
+        $this->authorize('restore', [Role::class]);
+
         $user = Auth::guard('api')->user();
 
         $role = Role::onlyRelated($user)->onlyTrashed()->findOrFail($id);
@@ -150,13 +176,27 @@ class RoleController extends Controller
      *
      * @param string $id
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
+        $this->authorize('forceDelete', [Role::class]);
+
         $user = Auth::guard('api')->user();
 
         $role = Role::onlyRelated($user)->onlyTrashed()->findOrFail($id);
+
+        // ...
+
+        $dependants = User::withTrashed()->where('role_id', $id)->count();
+
+        if ($dependants) {
+            return response(['message' => "Can't delete non-orphaned role."], 400);
+        }
+
+        // ...
 
         $role->forceDelete();
 
@@ -174,6 +214,8 @@ class RoleController extends Controller
      */
     public function users($id)
     {
+        $this->authorize('viewAny', [User::class]);
+
         $user = Auth::guard('api')->user();
 
         $role = Role::onlyRelated($user)->findOrFail($id);
@@ -186,10 +228,14 @@ class RoleController extends Controller
      *
      * @param string $id
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function permissions($id)
     {
+        $this->authorize('viewPermissions', [Role::class]);
+
         $user = Auth::guard('api')->user();
 
         $role = Role::onlyRelated($user)->findOrFail($id);
@@ -202,10 +248,14 @@ class RoleController extends Controller
      *
      * @param string $id
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
-    public function permissions_granted($id)
+    public function permissionsGranted($id)
     {
+        $this->authorize('viewPermissions', [Role::class]);
+
         $user = Auth::guard('api')->user();
 
         $role = Role::onlyRelated($user)->findOrFail($id);
@@ -244,10 +294,15 @@ class RoleController extends Controller
      * @param \Illuminate\Http\Request $request
      * @param string                   $id
      *
+     * @throws ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
-    public function sync_permissions(Request $request, $id)
+    public function syncPermissions(Request $request, $id)
     {
+        $this->authorize('assignPermissions', [Permission::class]);
+
         $role = Role::findOrFail($id);
 
         $this->validate($request, [
